// Backed/interfaces/contact.interface.ts
// export interface Contact {
//     name: string;
//     email: string;
//     subject: string;
//     message: string;
//     createdAt?: Date;
// }
export interface IEmail {
    name: string;
    email: string;
    subject: string;
    message: string;
    createdAt?: Date;
  }
  
  export interface ISubscriptionEmail {
    email: string;
  }